Fix MatrixMixer in Mixer doesn't work
The root cause of the issue is broken engine graph.
The reason that engine graph's get broken is a combination
of two factors

- The order of how we make connections
- Breaking/adding connection multiple times to a node

When using connect to connection points API and providing multiple
connection points, AVAudioEngine will internally create multi splitter
node (this can be verified by printing avEngine).

This splitter node is more sensitive it seems then explicit nodes and is
more likely to introduce broken graph.

Experimentally, I discovered that we need to:

1. Make connections in order from source node to destination nodes.
   The existing logic was doing the opposite
2. Do not connect one node multiple times. The existing logic was recursively enumerating
connections and the same node could be encountered multiple times in
different chain paths resulting in breaking and establishing existing
connections.

All tests pass when using optimised strategy, but I thought it is better
to keep existing behaviour the default and if new strategy proves good enough,
we can easily make it a default.